+++ src/js/controller/SubmitHandler.js
@@ -1,5 +1,5 @@
-import { parseLiteral } from "./CodeParser.js";
-import { controllEventLoop } from "./EventLoop.js";
+import { parseLiteral } from "../model/CodeParser.js";
+import { controllEventLoop } from "../view/EventLoop.js";
 
 export async function handleSubmit() {
   const runButton = document.querySelector(".submit-btn");
@@ -9,4 +9,4 @@ export async function handleSubmit() {
     const callbacks = await parseLiteral(code).then((resolve) => resolve);
     controllEventLoop(callbacks);
   });
-}
+};
